Add test for calculating kernel matrix based on demo data
