+++ tests/stein_thinning/test_thinning.py
@@ -7,21 +7,6 @@ import pytest
 from stein_thinning.thinning import thin
 
 
-@pytest.fixture
-def demo_data_dir():
-    return Path('stein_thinning') / 'sample_chains' / 'gmm'
-
-
-@pytest.fixture
-def demo_smp(demo_data_dir):
-    return np.genfromtxt(demo_data_dir / 'smp.csv', delimiter=',')
-
-
-@pytest.fixture
-def demo_scr(demo_data_dir):
-    return np.genfromtxt(demo_data_dir / 'scr.csv', delimiter=',')
-
-
 def test_thin(demo_smp, demo_scr):
     idx = thin(demo_smp, demo_scr, 40)
     expected = np.array([
@@ -30,4 +15,4 @@ def test_thin(demo_smp, demo_scr):
         148, 260, 296, 208,  79, 430, 369, 363, 462, 393, 321, 460, 373,
         114
     ])
-    np.testing.assert_array_equal(idx, expected)
+    np.testing.assert_array_almost_equal(idx, expected)
